add: firestore

+++ repository/post.repository.go
@@ -67,7 +67,7 @@ func (*repo) FindAll() ([]entity.Post, error) {
 			return nil, err
 		}
 		post := entity.Post{
-			ID:    doc.Data()["ID"].(int),
+			ID:    doc.Data()["ID"].(int64),
 			Title: doc.Data()["Title"].(string),
 			Text:  doc.Data()["Text"].(string),
 		}
